+++ CHANGELOG.md
@@ -11,6 +11,7 @@ e este projeto adota [Versionamento Semântico](https://semver.org/lang/pt-BR/).
 - Implementacao do componentes de SelectField.
 - Implementação do componente TextFieldBirthDateWithAge.
 - Implementação das funcoes de validacao e formatacao de CPF e CEP.
+- Implementação do componente TextFieldCPFValidate.
 
 ## [1.0.0] - 2025-12-28
 - Criação do package